Finished genome to pytorch nn conversion

+++ NeuroEvo/NeuralNetwork/Layer.py
@@ -30,5 +30,5 @@ class Layer(nn.Module):
     def forward(self, x):
         print(x)
         print(self.weights.t())
-        w_times_x = torch.matmul(x, self.weights.t())
+        w_times_x = torch.matmul(x.float(), self.weights.t())
         return torch.add(w_times_x, self.bias) # w times x + b
